+++ lib/scorched/static.rb
@@ -1,14 +1,18 @@
 module Scorched
   class Static
     def initialize(app, dir = 'public')
-      @app = app
-      @file_server = Rack::File.new(dir)
+      @app, @dir = app, dir
     end
 
     def call(env)
-      @file_server.call(env)
-      response = @file_server.call(env)
-      response[0] >= 400 ? @app.call(env) : response 
+      response = file_server.call(env)
+      response[0] >= 400 ? @app.call(env) : response
+    end
+
+  protected
+
+    def file_server
+      @file_server ||= Rack::File.new(@dir)
     end
   end
 end