+++ CHANGES.md
@@ -5,6 +5,9 @@ _Note that Scorched is yet to reach a v1.0 release. This means breaking changes 
 
 ### v0.24
 *  Query string is now preserved when stripping trailing slashes with `config[:strip_trailing_slash] = :redirect`, e.g. `/search/?query=cats` now becomes `/search?query=cats` instead of just `/search`.
+* `absolute` method now returns as-is anything not starting with a forward slash, e.g. `absolute("./view") => "./view"`
+* Improved performance by caching the Rack::Builder instance for each controller.
+* Improved performance by caching data associated with Scorched::Options#to_hash.
 
 ### v0.23
 * Now using _scorched-accept_ for accept header parsing and logic, which fixes issues with the `:media_type` condition.